Fixes issue #4
Recent releases of CubeMX (from 4.14 and higher) allow to generate the
SW4STM32 project in the root folder. This means that project files are
stored in the root of the CubeMX project, but this is the same behavior
for TrueSTUDIO project. So we need to check if the project is generated
for the SW4STM32 toolchain by playing with the content of .cproject file